+++ tests/test_agents/test_positionfac.h
@@ -9,7 +9,7 @@ namespace cyclus {
 /// This is the simplest possible PositionFac, for testing
 class TestPositionFac : public cyclus::toolkit::Position {
  public:
-  TestPositionFac(long latitude, long longitude)
+  TestPositionFac(double latitude, double longitude)
       : cyclus::toolkit::Position(0, 0) {}
   virtual ~TestPositionFac() {}
   // virtual cyclus::Agent* Clone() { return new TestPositionFac(context()); }
